feat : 라인 위치 함수 분리
