+++ frontend/src/components/PrCommentHeader.module.css
@@ -2,6 +2,7 @@
   display: flex;
   gap: 2rem;
   margin-top: 1.5rem;
+  flex-wrap: wrap;
 }
 .prCommentDesc div {
   background: white;
