+++ frontend/src/pages/PRCommentPage.module.css
@@ -115,7 +115,7 @@
   background: white;
   padding: 1rem 1.5rem;
   border-radius: 21px;
-  border: 1px solid lightgray;
+  border: 1px solid rgb(25, 19, 19);
   box-shadow: 5px 5px 5px #d3d3d37a;
   width: 100%;
 }
@@ -153,6 +153,6 @@ pre > div {
 }
 .codeLineWrapper .commentBody {
   width: 60%;
+  border-radius: 0;
   padding: 1rem;
-  margin-bottom: 1rem;
 }
